Move smt2irep testing utilities into `unit/testing-utils`
So they can be reused in multiple test `.cpp` files.
