+++ unit/testing-utils/module_dependencies.txt
@@ -3,5 +3,6 @@ ansi-c
 catch
 goto-programs
 langapi
+solvers/smt2
 testing-utils
 util
